size

+++ 0x00-hello_world/6-size.c
@@ -7,10 +7,10 @@
 */
 int main(void)
 {
-	printf("Size of a char: %i byte(s)\n", (unsigned)sizeof(char));
-	printf("Size of a int: %i byte(s)\n", (unsigned)sizeof(int));
-	printf("Size of a long int: %i byte(s)\n", (unsigned)sizeof(long int));
-	printf("Size of a long long: %i byte(s)\n", (unsigned)sizeof(long long));
-	printf("Size of a float: %i byte(s)\n", (unsigned)sizeof(float));
+	printf("Size of a char: %lu byte(s)\n", sizeof(char));
+	printf("Size of a int: %lu byte(s)\n", sizeof(int));
+	printf("Size of a long int: %lu byte(s)\n", sizeof(long int));
+	printf("Size of a long long: %lu byte(s)\n", sizeof(long long));
+	printf("Size of a float: %lu byte(s)\n", sizeof(float));
 	return (0);
 }
